# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Tasker",
        "description": "This is a sample Petstore server.  You can find\nout more about Swagger at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/attachment/{id}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Remove an attachment",
                "operationId": "App\\Http\\Controllers\\AttachmentController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of attachment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return success"
                    }
                }
            }
        },
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get list of tasks",
                "operationId": "App\\Http\\Controllers\\TaskController::index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list of tasks"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create a new task",
                "operationId": "App\\Http\\Controllers\\TaskController::store",
                "requestBody": {
                    "description": "Create data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Task's name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Task's description",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Task's price",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "deadline": {
                                        "description": "Task's deadline",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "online_or_phone": {
                                        "description": "Task's mode",
                                        "type": "boolean"
                                    },
                                    "specified_times": {
                                        "description": "Task's specified times",
                                        "properties": {
                                            "morning": {
                                                "type": "boolean"
                                            },
                                            "midday": {
                                                "type": "boolean"
                                            },
                                            "afternoon": {
                                                "type": "boolean"
                                            },
                                            "evening": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "default_location": {
                                        "description": "Task's default location",
                                        "properties": {
                                            "display_name": {
                                                "type": "string"
                                            },
                                            "longtitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create a task successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/tasks/{slug}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get a single task by slug",
                "operationId": "App\\Http\\Controllers\\TaskController::show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a single task by slug"
                    }
                }
            }
        },
        "/users/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload current user's avatar",
                "operationId": "App\\Http\\Controllers\\UserAvatarController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "image/jpeg": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "avatar": {
                                        "description": "avatar file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return list of tasks"
                    }
                }
            }
        },
        "/users/{slug}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Find user by slug",
                "description": "Returns a single user",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of user to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Slug supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user",
                "description": "Returns current user",
                "operationId": "App\\Http\\Controllers\\UserController::me",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "User not authenticated"
                    }
                }
            }
        }
    }
}